+++ nextgisweb/svg_marker_library/util.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from __future__ import division, absolute_import, print_function, unicode_literals
+from nextgisweb.i18n import trstring_factory
+
+COMP_ID = 'svg_marker_library'
+_ = trstring_factory(COMP_ID)